Moved main in his own file. Included user-input.hpp in files instead of all includes Corrected some compilation error
/!\ Not working Yet /!\
