+++ include/user-input/user-input.hpp
@@ -12,6 +12,8 @@
 #include "user-input/listeners.hpp"
 #include "user-input/window.hpp"
 
+//class Listeners;
+
 class userInput {
 public:
   static userInput &get();
