+++ source/user-input/listeners/registryListener.cpp
@@ -2,7 +2,7 @@
 
 registryListener::registryListener(std::map<std::string, var_listener> &listeners) {
   static struct wl_registry_listener registry_listener =
-          {&registryListener::egistry_add_object,
+          {&registryListener::registry_add_object,
            &registryListener::registry_remove_object};
   listeners.emplace("registry", registry_listener);
 }
@@ -16,10 +16,10 @@ void registryListener::registry_add_object (void *data, struct wl_registry *regi
 	}
 	else if (!strcmp(interface,"wl_seat")) {
 		userInput::get().seat = static_cast<wl_seat *>(wl_registry_bind(registry, name, &wl_seat_interface, 1));
-		wl_seat_add_listener(userInput::get().seat, &seat_listener, NULL);
+		wl_seat_add_listener(userInput::get().seat, &(userInput::get().listeners->getListener("seat")), NULL);
 	}
 }
 
-void registryListener::egistry_remove_object (void *data, struct wl_registry *registry, uint32_t name) {
+void registryListener::registry_remove_object (void *data, struct wl_registry *registry, uint32_t name) {
 
 }
