+++ source/user-input/listeners/seatListener.cpp
@@ -2,16 +2,16 @@
 
 seatListener::seatListener(std::map<std::string, var_listener> &listeners) {
   static struct wl_seat_listener seat_listener = {&seatListener::seat_capabilities};
-  listeners.empalce("seat", seat_listener);
+  listeners.emplace("seat", seat_listener);
 }
 
 void seat_capabilities (void *data, struct wl_seat *seat, uint32_t capabilities) {
 	if (capabilities & WL_SEAT_CAPABILITY_POINTER) {
 		struct wl_pointer *pointer = wl_seat_get_pointer(seat);
-		wl_pointer_add_listener(pointer, &pointer_listener, NULL);
+		wl_pointer_add_listener(pointer, &(userInput::get().listeners->getListener("pointer")), NULL);
 	}
 	if (capabilities & WL_SEAT_CAPABILITY_KEYBOARD) {
 		struct wl_keyboard *keyboard = wl_seat_get_keyboard(seat);
-		wl_keyboard_add_listener(keyboard, &keyboard_listener, NULL);
+		wl_keyboard_add_listener(keyboard, &(userInput::get().listeners->getListener("keyboard")), NULL);
 	}
 }
